微信端活动修改

+++ src/main/java/com/activity/controller/WechatActivityController.java
@@ -1,8 +1,27 @@
 package com.activity.controller;
 
-import com.activity.model.*;
+import com.activity.model.Activity;
+import com.activity.model.ActivityCourse;
+import com.activity.model.ActivityDescription;
+import com.activity.model.ActivityDistrict;
+import com.activity.model.ActivityEnroll;
+import com.activity.model.ActivityTag;
+import com.activity.model.ActivityThumbup;
+import com.activity.model.ActivityWatched;
+import com.activity.model.Adsense;
+import com.activity.model.UploadFile;
+import com.activity.model.Users;
+import com.activity.model.WechatUser;
 import com.activity.pojo.WechatPojo;
-import com.activity.service.*;
+import com.activity.service.ActivityDistrictService;
+import com.activity.service.ActivityEnrollService;
+import com.activity.service.ActivityService;
+import com.activity.service.ActivityThumbupService;
+import com.activity.service.ActivityWatchedService;
+import com.activity.service.AdsenseService;
+import com.activity.service.UploadFileService;
+import com.activity.service.UsersService;
+import com.activity.service.WechatUserService;
 import com.activity.utils.Constants;
 import com.activity.utils.DateUtils;
 import com.activity.utils.RestEntity;
@@ -12,7 +31,12 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.util.ObjectUtils;
 import org.springframework.util.StringUtils;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.HashMap;
@@ -57,19 +81,22 @@ public class WechatActivityController {
      * Created by ky.bai on 2018-03-01 14:49
      *
      * @param openid 用户微信openid
+     * @param state  用户微信标识(有此字段才会加载我的信息), 分享或复制链接不会有state参数[state=INDEX], 带此参数的的入口：微信按钮和授权和扫码签到
      */
     @RequestMapping(value = "", method = RequestMethod.GET)
-    public String list(@RequestParam(required = false) String openid, Model model) {
-        model.addAttribute("openid", openid);
+    public String list(@RequestParam(required = false) String openid, @RequestParam(required = false) String state, Model model) {
         model.addAttribute("districts", activityDistrictService.selectList(new ActivityDistrict(Boolean.TRUE)));
         //轮播图
         model.addAttribute("imgList", adsenseService.selectList(new Adsense(Constants.IMAGE_TYPE_ROLLING, Boolean.TRUE)));
-        if (!StringUtils.isEmpty(openid)) {
+        if (!StringUtils.isEmpty(openid) && !StringUtils.isEmpty(state) && state.equals(Constants.WECHAT_STATE_INDEX)) {
             model.addAttribute("mylist", activityService.selectUserWechatList(new WechatPojo(openid, Boolean.TRUE)));
             //我的信息与积分
             WechatUser wechatUser = wechatUserService.findByOpenid(openid);
             model.addAttribute("user", usersService.selectUserScore(wechatUser.getUserId()));
+            model.addAttribute("hasAuth", Boolean.TRUE);
+            model.addAttribute("openid", openid);
         }
+        model.addAttribute("hasAuth", Boolean.FALSE);
         return "wechat/list";
     }
 
@@ -85,7 +112,7 @@ public class WechatActivityController {
         if (activity == null) {
             return "wechat/detail";
         }
-        model.addAttribute("openid", openid);
+        //model.addAttribute("openid", openid);
         //活动与活动简介
         model.addAttribute("activity", activity);
         model.addAttribute("file", uploadFileService.selectOne(activity.getUploadFileId()));
@@ -122,6 +149,7 @@ public class WechatActivityController {
      * Created by ky.bai on 2018-03-01 11:30
      *
      * @param pojo 微信端请求参数
+     *
      * @return 微信端活动列表
      */
     @RequestMapping(value = "/list", method = RequestMethod.POST)
@@ -140,6 +168,7 @@ public class WechatActivityController {
      * Created by ky.bai on 2018-03-02 09:26
      *
      * @param pojo 活动编号, 用户微信openid
+     *
      * @return 活动点赞记录
      */
     @RequestMapping(value = "/thumbup", method = RequestMethod.POST)
@@ -147,16 +176,19 @@ public class WechatActivityController {
     public ResponseEntity thumbup(@RequestBody WechatPojo pojo) {
         String openid = pojo.getOpenid();
         Integer activityId = pojo.getActivityId();
-        WechatUser user = wechatUserService.findByOpenid(openid);
-        List<ActivityThumbup> thumbupList = activityThumbupService.selectList(new ActivityThumbup(activityId, user.getUserId()));
-        if (!ObjectUtils.isEmpty(thumbupList)) {
-            ActivityThumbup thumbup = thumbupList.get(0);
-            activityThumbupService.delete(thumbup);
-        } else {
-            activityThumbupService.insert(new ActivityThumbup(activityId, user.getUserId()));
+        if (!StringUtils.isEmpty(openid) && activityId != null) {
+            WechatUser user = wechatUserService.findByOpenid(openid);
+            List<ActivityThumbup> thumbupList = activityThumbupService.selectList(new ActivityThumbup(activityId, user.getUserId()));
+            if (!ObjectUtils.isEmpty(thumbupList)) {
+                ActivityThumbup thumbup = thumbupList.get(0);
+                activityThumbupService.delete(thumbup);
+            } else {
+                activityThumbupService.insert(new ActivityThumbup(activityId, user.getUserId()));
+            }
+            return ResponseEntity.ok(new RestEntity(200, Constants.OPERATOR_SUCCESS, Boolean.TRUE));
         }
 
-        return ResponseEntity.ok(new RestEntity(200, Constants.OPERATOR_SUCCESS, Boolean.TRUE));
+        return ResponseEntity.ok(new RestEntity(100, Constants.OPERATOR_SUCCESS, Boolean.FALSE));
     }
 
     /**
@@ -164,17 +196,19 @@ public class WechatActivityController {
      *
      * @param activityId 活动编号
      * @param openid     微信用户openid
+     *
      * @return 活动报名页
      */
     @RequestMapping(value = "/enroll/{activityId}", method = RequestMethod.GET)
     public String activityEnroll(@PathVariable Integer activityId, @RequestParam String openid, Model model) {
-        model.addAttribute("openid", openid);
-        WechatUser user = wechatUserService.findByOpenid(openid);
-        model.addAttribute("userId", user.getUserId());
         Activity activity = activityService.selectOne(activityId);
         model.addAttribute("activity", activity);
         model.addAttribute("courses", activityService.selectCourseList(new ActivityCourse(activityId, Boolean.TRUE)));
         model.addAttribute("tag", activityService.selectTag(new ActivityTag(activityId)));
+        if (!StringUtils.isEmpty(openid)) {
+            WechatUser user = wechatUserService.findByOpenid(openid);
+            model.addAttribute("userId", user.getUserId());
+        }
         return "wechat/signUp";
     }
 
@@ -182,27 +216,24 @@ public class WechatActivityController {
      * Created by ky.bai on 2018-03-02 13:56
      *
      * @param record 报名信息
+     *
      * @return 活动报名结果
      */
     @RequestMapping(value = "/enroll", method = RequestMethod.POST)
     @ResponseBody
     public ResponseEntity saveEnroll(@RequestBody ActivityEnroll record) {
         Activity activity = activityService.selectOne(record.getActivityId());
-        if (activity == null)
-            return ResponseEntity.ok(new RestEntity(100, Constants.ENROLL_RESULT_NOT_FOUND, Boolean.FALSE));
+        if (activity == null) return ResponseEntity.ok(new RestEntity(100, Constants.ENROLL_RESULT_NOT_FOUND, Boolean.FALSE));
 
         Users user = usersService.selectOne(record.getUserId());
-        if (user != null && !user.getActive())
-            return ResponseEntity.ok(new RestEntity(100, Constants.ENROLL_RESULT_WAS_DISABLED, Boolean.FALSE));
+        if (user != null && !user.getActive()) return ResponseEntity.ok(new RestEntity(100, Constants.ENROLL_RESULT_WAS_DISABLED, Boolean.FALSE));
 
         //判断当前活动是否在开始与结束时间范围内
-        if (activity.getEndTime().getTime() < DateUtils.getCurrentTimestamp().getTime())
-            return ResponseEntity.ok(new RestEntity(100, Constants.ENROLL_RESULT_WAS_END, Boolean.FALSE));
+        if (activity.getEndTime().getTime() < DateUtils.getCurrentTimestamp().getTime()) return ResponseEntity.ok(new RestEntity(100, Constants.ENROLL_RESULT_WAS_END, Boolean.FALSE));
 
         //判断是否已经报名
         List<ActivityEnroll> enrolls = activityEnrollService.selectList(new ActivityEnroll(activity.getId(), record.getUserId(), Boolean.TRUE));
-        if (!ObjectUtils.isEmpty(enrolls))
-            return ResponseEntity.ok(new RestEntity(100, Constants.ENROLL_RESULT_WAS_SIGN, Boolean.FALSE));
+        if (!ObjectUtils.isEmpty(enrolls)) return ResponseEntity.ok(new RestEntity(100, Constants.ENROLL_RESULT_WAS_SIGN, Boolean.FALSE));
 
         //判断已报名人数是否已满
         List<ActivityEnroll> enrollList = activityEnrollService.selectList(new ActivityEnroll(activity.getId(), Boolean.TRUE));
@@ -217,6 +248,7 @@ public class WechatActivityController {
      * Created by ky.bai on 2018-03-02 13:55
      *
      * @param file 上传文件
+     *
      * @return 上传文件编号
      */
     @RequestMapping(value = "/upload", method = RequestMethod.POST)
@@ -229,7 +261,6 @@ public class WechatActivityController {
     @RequestMapping(value = "/enroll/{activityId}/success", method = RequestMethod.GET)
     public String enterSuccess(@PathVariable Integer activityId, @RequestParam String openid, Model model) {
         model.addAttribute("activityId", activityId);
-        model.addAttribute("openid", openid);
         return "wechat/enterSuccess";
     }
 
@@ -239,11 +270,11 @@ public class WechatActivityController {
      * @param activity    活动
      * @param userId      用户编号
      * @param enrollTotal 活动已报名人数
+     *
      * @return 活动状态: --我要报名，--已报名，--名额已满，--已结束
      */
     private String getActivityStatus(Activity activity, Integer userId, int enrollTotal) {
         if (userId == null) return Constants.ENROLL_WECHAT;
-        Users user = usersService.selectOne(userId);
 
         List<ActivityEnroll> enrolls = activityEnrollService.selectList(new ActivityEnroll(activity.getId(), userId, Boolean.TRUE));
         long currentTime = DateUtils.getCurrentTimestamp().getTime();
